refactor: changed types of some columns

+++ NKRY-API/Domain/Entities/EmployeeFinance.cs
@@ -14,8 +14,8 @@ namespace NKRY_API.Domain.Entities
         // can be changed by the admin only
         public int DeliveryRate { get; set; }
         [Required]
-        public float BaseSalary { get; set; }
-        public float? TotalSalary { get; set; }
+        public double BaseSalary { get; set; }
+        public double? TotalSalary { get; set; }
         public string? DueMonth { get; set; }
         public DateTime? DueDate { get; set; }
         public Guid EmployeeId { get; set; }
